auto: issue #39 Split backend/graph/session_manager.py (802 lines)
